pyproject.toml
[tool.poetry]
name = "comic-users"
version = "0.1.0"
description = "Crud for user data"
authors = ["Fausto Ramos <fausto.ramosr92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
pymongo = "^4.1.1"
dnspython = "^2.2.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"
black = "^22.3.0"
flake8 = "^4.0.1"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
